test(abw): split image route tests
